--- src/integrations/skadi/api/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export { SkadiApiClient } from './clients';
