+++ src/schema/posts.ts
@@ -146,7 +146,7 @@ import {
   UserTransactionStatus,
   UserTransactionType,
 } from '../entity/user/UserTransaction';
-import { skadiApiClient } from '../integrations/skadi/api/clients';
+import { skadiApiClientV1 } from '../integrations/skadi/api/v1/clients';
 import {
   validatePostBoostPermissions,
   checkPostAlreadyBoosted,
@@ -2231,7 +2231,7 @@ export const resolvers: IResolvers<unknown, BaseContext> = traceResolvers<
       const post = await validatePostBoostPermissions(ctx, postId);
       checkPostAlreadyBoosted(post);
 
-      const { users } = await skadiApiClient.estimatePostBoostReach({
+      const { users } = await skadiApiClientV1.estimatePostBoostReach({
         postId,
         userId: ctx.userId,
       });
@@ -2249,7 +2249,7 @@ export const resolvers: IResolvers<unknown, BaseContext> = traceResolvers<
       validateCampaignArgs({ budget, duration });
 
       const { minImpressions, maxImpressions } =
-        await skadiApiClient.estimatePostBoostReachDaily({
+        await skadiApiClientV1.estimatePostBoostReachDaily({
           postId,
           userId: ctx.userId,
           budget: coresToUsd(budget),
@@ -2270,7 +2270,7 @@ export const resolvers: IResolvers<unknown, BaseContext> = traceResolvers<
       { id }: { id: string },
       ctx: Context,
     ): Promise<GQLBoostedPost> => {
-      const campaign = await skadiApiClient.getCampaignById({
+      const campaign = await skadiApiClientV1.getCampaignById({
         campaignId: id,
         userId: ctx.userId!,
       });
@@ -2309,7 +2309,7 @@ export const resolvers: IResolvers<unknown, BaseContext> = traceResolvers<
         (nodeSize) => nodeSize === first,
         (_, i) => offsetToCursor(offset + i + 1),
         async () => {
-          const campaigns = await skadiApiClient.getCampaigns({
+          const campaigns = await skadiApiClientV1.getCampaigns({
             userId,
             offset,
             limit: first!,
@@ -2852,7 +2852,7 @@ export const resolvers: IResolvers<unknown, BaseContext> = traceResolvers<
       const total = budget * duration;
 
       const request = await ctx.con.transaction(async (entityManager) => {
-        const { campaignId } = await skadiApiClient.startPostCampaign({
+        const { campaignId } = await skadiApiClientV1.startPostCampaign({
           postId,
           durationInDays: duration,
           budget: coresToUsd(budget),
@@ -2933,7 +2933,7 @@ export const resolvers: IResolvers<unknown, BaseContext> = traceResolvers<
       }
 
       const result = await ctx.con.transaction(async (entityManager) => {
-        const { currentBudget } = await skadiApiClient.cancelPostCampaign({
+        const { currentBudget } = await skadiApiClientV1.cancelPostCampaign({
           campaignId,
           userId: ctx.userId,
         });
